Create IterativeFactorial.kt
This is the iterative version of the factorial problem. I would like to note I cannot use the classic JavaScript for loop for Kotlin as it appears currently. THe best option for me is to declare the conditional variable of the loop outisde of a while looop and then make my condition inside of said while loop with it's conditional logic within itself.